requirements.txt
boto3
gradio
httpx
pydantic
pytest
python-dotenv
